requirements.txt
aggdraw==1.3.15
attrs==22.1.0
docopt==0.6.2
imageio==2.22.2
networkx==2.8.7
numpy==1.23.4
packaging==21.3
Pillow==9.2.0
psd-tools==1.9.23
pyparsing==3.0.9
PyWavelets==1.4.1
scikit-image==0.19.3
scipy==1.9.3
tifffile==2022.10.10
